Move HealthCheckHandlerTest to correct package Create a test for findUnhealthyCores, factoring it into a separate method Rename param as 'requireHealthyCores'
